コンポーネント化しよう（商品在庫）
